requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-extensions==2.2.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
